rename : 퀴즈 페이지들 공용 컴포넌트 위치 변경
