+++ src/components/quiz/svg/ArrowRightIcon.tsx
