+++ src/components/quiz/svg/CheckedIcon.tsx
