+++ src/components/quiz/svg/ChumSungDaeIcon.tsx
