+++ src/components/quiz/svg/CrossedIcon.tsx
