+++ src/components/quiz/svg/GagsiMaskIcon.tsx
