+++ src/components/quiz/svg/GyeongBokGungIcon.tsx
